BuildRequires: polkit for tests
The polkit-devel build dependency pulls in polkit-libs for linking against
the library, but (by design) not the polkit daemon.  However, since 2.0.6,
the fu-polkit-test requires polkit itself to be present, otherwise it fails:

ERROR:../src/fu-polkit-test.c:19:fu_polkit_agent_func: assertion failed (error == NULL): missing pkttyagent (FwupdError, 10)

When building on Fedora, the ModemManager-devel dependency pulls in
ModemManager (as there is no -libs subpackage) which requires polkit.
However, that dependency is disabled for RHEL and ELN, in which case
nothing pulls in polkit and the test fails.  An explicit dependency is
therefore necessary.

https://github.com/fwupd/fwupd/commit/2f31915e925af0865aaf2a56f227834cb4add2e4